requirements.txt
python-dateutil==2.8.2
retrying==1.3.3
pymysql==1.0.2
redis==4.3.4
pyyaml==6.0
numpy==1.23.3
pandas==1.4.4
requests==2.28.1
apscheduler==3.9.1
